moving things around and start working on the hoisting algorithem
