+++ src/extensions/dependency-resolver/manifest/deduping/index.ts
@@ -1,0 +1,1 @@
+export { dedupeDependencies, DedupedDependencies } from './dedupe-dependencies';
